Remove references to binary data in APIs and SDK
Quixstreams <1 had support specific binary parameters, however in quixstreams 2.x messages were designed to be less strict in schema.
With that, Binary messages are possible if serialization is provided and don't need special handling.
In some deprecated Quix APIs there was support for these parameters but it is no longer present either.